+++ internal/domain/repository/authentications.go
@@ -2,7 +2,7 @@
 package repository
 
 type Authentications interface {
-	Create(identify, token string) (int64, error)
-	FindToken(identify string) (int64, string, error)
+	Create(identify, token string) (int, error)
+	FindToken(identify string) (int, string, error)
 	Close()
 }
