+++ internal/domain/repository/individuals.go
@@ -6,8 +6,8 @@ import (
 )
 
 type Individuals interface {
-	Create(name string, userId, profileId int64) (int64, error)
-	Find(id int64) (*entity.Individual, error)
-	FindByUserId(userId int64) (*[]entity.Individual, error)
+	Create(name string, userId, profileId int) (int, error)
+	Find(id int) (*entity.Individual, error)
+	FindByUserId(userId int) (*[]entity.Individual, error)
 	Close()
 }
